열여섯번째 커밋 : JWT 인증 토큰 만료 및 오류에 대한 예외 처리 로직 추가

+++ controller/LikeController.js
@@ -1,36 +1,62 @@
+const ensureAuthorization = require('../auth');
+const jwt = require('jsonwebtoken');
 const conn = require('../mariadb');
 const {StatusCodes} = require('http-status-codes');
 
-const addLike = (req, res) => {
-    const {id} = req.params;
-    const {user_id} = req.body;
 
-    let sql = `INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)`;
-    let values = [user_id, id];
-    conn.query(sql, values,
-        (err, results) => {
-            if(err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
-            }
-            return res.status(StatusCodes.OK).json(results);
-        })
+const addLike = (req, res) => {
+    const book_id = req.params.id;
+
+    let authorization = ensureAuthorization(req, res);
+    console.log("Authorizaiton result:" ,authorization);
+
+    if(authorization instanceof jwt.TokenExpiredError) {
+            return res.status(StatusCodes.UNAUTHORIZED).json({
+                "message" : "로그인 세션이 만료되었습니다. 다시 로그인하세요."
+            });
+        } else if (authorization instanceof jwt.JsonWebTokenError){
+            return res.status(StatusCodes.BAD_REQUEST).json({
+                "message" : "잘못된 토큰입니다."
+            });
+        } else {
+            let sql = `INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)`;
+            let values = [authorization.id, book_id];
+            conn.query(sql, values,
+                (err, results) => {
+                    if(err) {
+                        console.log(err);
+                        return res.status(StatusCodes.BAD_REQUEST).end();
+                    }
+                    return res.status(StatusCodes.OK).json(results);
+                })
+        }
 };
 
 const removeLike = (req, res) => {
-    const {id} = req.params;
-    const {user_id} = req.body;
+    const book_id  = req.params.id;
 
-    let sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?`;
-    let values = [user_id, id];
-    conn.query(sql, values,
-        (err, results) => {
-            if(err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
-            }
-            return res.status(StatusCodes.OK).json(results);
-        })
+    let authorization = ensureAuthorization(req, res);
+
+    if(authorization instanceof jwt.TokenExpiredError) {
+            return res.status(StatusCodes.UNAUTHORIZED).json({
+                "message" : "로그인 세션이 만료되었습니다. 다시 로그인하세요."
+            });
+        } else if (authorization instanceof jwt.JsonWebTokenError){
+            return res.status(StatusCodes.BAD_REQUEST).json({
+                "message" : "잘못된 토큰입니다."
+            });
+        } else {
+            let sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?`;
+            let values = [authorization.id, book_id];
+            conn.query(sql, values,
+                (err, results) => {
+                    if(err) {
+                        console.log(err);
+                        return res.status(StatusCodes.BAD_REQUEST).end();
+                    }
+                    return res.status(StatusCodes.OK).json(results);
+                })
+        }
 };
 
 module.exports = {
